Version 1.1.7 final Shift and Click can be used to select a point DetailView is set to the prediction when no point is detected or it does not move if no prediction is possible. ZoomLevel of Detailview is not reset when changing frames All DockWidgets save the position and size (Plot, 3D View etc.) Closable Dockwidgets can also be closed by using the close-button in the title 3 advances forward to the last tracked frame. Shift + 3 advances backward to the last tracked frame.

+++ src/processing/MarkerTracking.h
@@ -44,7 +44,7 @@ namespace xma{
 		double x_to;
 		double y_to;
 
-		double size;
+		int size;
 		int searchArea;
 		int maxPenalty;
